Simplify home page, fixes from code review
