+++ packages/back-end/src/app.ts
@@ -482,10 +482,6 @@ app.get("/feature/:id", featuresController.getFeatureById);
 app.post("/feature", featuresController.postFeatures);
 app.put("/feature/:id", featuresController.putFeature);
 app.delete("/feature/:id", featuresController.deleteFeatureById);
-app.get(
-  "/feature/frequency/month/:num",
-  featuresController.getFeaturesFrequencyMonth
-);
 app.get("/usage/features", featuresController.getRealtimeUsage);
 
 // Data Sources
